+++ actix/src/fut/future/mod.rs
@@ -158,6 +158,9 @@ pub trait ActorFutureExt<A: Actor>: ActorFuture<A> {
         Timeout::new(self, timeout)
     }
 
+    /// Wrap the future in a Box, pinning it.
+    ///
+    /// A shortcut for [`Box::pin`](std::boxed::Box::pin)
     fn boxed_local(self) -> LocalBoxActorFuture<A, Self::Output>
     where
         Self: Sized + 'static,