+++ actix/src/address/mod.rs
@@ -372,11 +372,13 @@ where
 }
 
 impl<M> Clone for WeakRecipient<M>
-where M: Message + Send,
-M::Result: Send {
+where
+    M: Message + Send,
+    M::Result: Send,
+{
     fn clone(&self) -> Self {
         Self {
-            wtx : self.wtx.boxed()
+            wtx: self.wtx.boxed(),
         }
     }
 }