+++ actix/CHANGES.md
@@ -6,7 +6,7 @@
 * Add `fut::try_future::ActorTryFutureExt` trait with `map_ok`, `map_err` and `and_then` combinator. [#419] 
 
 ### Changed
-- Make `Context::new` public. #[491]
+- Make `Context::new` public. [#491]
 
 [#419]: https://github.com/actix/actix/pull/419
 [#491]: https://github.com/actix/actix/pull/491