+++ actix/CHANGES.md
@@ -3,13 +3,15 @@
 ## Unreleased - 2021-xx-xx
 ### Added
 * Add `fut::try_future::ActorTryFuture`. [#419]
-* Add `fut::try_future::ActorTryFutureExt` trait with `map_ok`, `map_err` and `and_then` combinator. [#419] 
+* Add `fut::try_future::ActorTryFutureExt` trait with `map_ok`, `map_err` and `and_then` combinator. [#419]
+* Add `fut::future::FutureExt::boxed_local` [#493] 
 
 ### Changed
 - Make `Context::new` public. [#491]
 
 [#419]: https://github.com/actix/actix/pull/419
 [#491]: https://github.com/actix/actix/pull/491
+[#493]: https://github.com/actix/actix/pull/493
 
 
 ## 0.11.1 - 2021-03-23