+++ actix/CHANGES.md
@@ -3,6 +3,8 @@
 ## Unreleased - 2021-xx-xx
 ### Removed
 - Removed `Resolver` actor [#451]
+- Implement `Clone` for `WeakRecipient` [#518]
+- 
 
 ## 0.12.0 - 2021-06-06
 ### Added